fix: 이너 클래스를 외부로 이동
