+++ src/main/java/im/juniq/racingcar/domain/RacingGame.java
@@ -45,22 +45,4 @@ public class RacingGame {
 				.map(car -> new Score(car.name().name(), car.position().position()))
 				.collect(Collectors.toList());
 	}
-
-	public class Score {
-		private String name;
-		private int position;
-
-		public Score(String name, int position) {
-			this.name = name;
-			this.position = position;
-		}
-
-		public String name() {
-			return name;
-		}
-
-		public int position() {
-			return position;
-		}
-	}
 }
