+++ src/test/java/im/juniq/racingcar/RacingGameTest.java
@@ -2,15 +2,15 @@ package im.juniq.racingcar;
 
 import static org.assertj.core.api.Assertions.*;
 
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
 import im.juniq.racingcar.domain.Car;
 import im.juniq.racingcar.domain.Name;
 import im.juniq.racingcar.domain.OnlyTrueMovingStrategy;
-import im.juniq.racingcar.domain.Position;
 import im.juniq.racingcar.domain.RacingGame;
-import im.juniq.racingcar.domain.RacingGame.Score;
 import im.juniq.racingcar.domain.RandomNumberMovingStrategy;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import im.juniq.racingcar.domain.Score;
 
 class RacingGameTest {
 	@DisplayName("자동차 n대를 생성한다")
